+++ src/lib.rs
@@ -30,22 +30,52 @@ mod error;
 pub use error::Error;
 use error::MyResult;
 mod post;
-use post::{EResult, EvaluatedTerm, RolledDie, TResult};
+use post::{EResult, EvaluatedTerm, RolledDie};
 pub use post::{ExpressionResult, FormatOptions};
 mod expose;
 #[cfg(feature = "thread_rng")]
 pub use expose::roll_tuples;
 pub use expose::tuple_vec;
-mod parse;
-use parse::{Die, Expr, Sign, Term};
+pub mod parse;
+use parse::{DiceTerm, Expr, Sign, Term};
 pub use parse::ParseError;
 pub mod builder;
 use builder::RollBuilder;
 mod display;
 pub mod prelude;
 pub mod util;
-
-fn roll_die_with<R>(a: &Die, rng: &mut R) -> Result<RolledDie, Error>
+#[derive(::thiserror::Error, Debug, Clone, Copy)]
+#[error("sum is too high for `i64`")]
+pub struct OverflowPositive;
+#[derive(::thiserror::Error, Debug, Clone, Copy)]
+#[error("sum is too low for `i64`")]
+pub struct OverflowNegative;
+#[derive(::thiserror::Error, Debug, Clone, Copy)]
+enum Overflow {
+    #[error(transparent)]
+    Positive(#[from] OverflowPositive),
+    #[error(transparent)]
+    Negative(#[from] OverflowNegative),
+}
+impl From<Overflow> for Error {
+    fn from(o: Overflow) -> Self {
+        match o {
+            Overflow::Positive(x) => Self::OverflowPositive(x),
+            Overflow::Negative(x) => Self::OverflowNegative(x),
+        }
+    }
+}
+impl ::core::ops::Neg for Overflow {
+    type Output = Overflow;
+    fn neg(self) -> Self::Output {
+        use Overflow::*;
+        match self {
+            Positive(x) => Negative(-x),
+            Negative(x) => Positive(-x),
+        }
+    }
+}
+fn roll_die_with<R>(a: &DiceTerm, rng: &mut R) -> Result<RolledDie, OverflowPositive>
 where
     R: Rng,
 {
@@ -55,8 +85,6 @@ where
             parts: (0..a.number).map(|_| 1).collect(),
             sign_part: Sign::Positive,
         })
-    } else if a.size < 1 {
-        Err(Error::InvalidDie)
     } else {
         let mut total: i64 = 0;
         let mut parts = Vec::new();
@@ -69,7 +97,7 @@ where
             } else {
                 random = rng.gen();
             }
-            total = total.checked_add(random).ok_or(Error::OverflowPositive)?;
+            total = total.checked_add(random).ok_or(OverflowPositive)?;
             parts.push(random);
         }
         Ok(RolledDie {
@@ -80,21 +108,16 @@ where
     }
 }
 
-fn eval_term_with<R>(a: &Expr, rng: &mut R) -> TResult
+fn eval_term_with<R>(a: &Expr, rng: &mut R) -> Result<EvaluatedTerm, Overflow>
 where
     R: Rng,
 {
-    let t = match a.term {
-        Term::Die(x) => roll_die_with(&x, rng).into(),
+    let t: MyResult<_, Overflow> = match a.term {
+        Term::Dice(x) => roll_die_with(&x, rng).into(),
         Term::Constant(x) => MyResult::Ok(EvaluatedTerm::Constant(x)),
     };
-    // This can't trigger a panic with any inputs as of now,
-    // but that may change. Switch to checked negation at some point.
-    // This currently cannot panic because the maximum permitted
-    // input is the `i64` positive max.
-    // This could panic if `t` were the `i64` negative max.
-    // See the `parse` module for the implementation of `Sign::mul`.
-    // I suggest providing a `Sign::checked_mul` method for this usage.
+    // No positive number can overflow via negation.
+    // Since terms are purely positive, this will never overflow.
     (a.sign * t).into()
 }
 
@@ -116,7 +139,7 @@ where
 ///   - Nonsense input
 #[cfg(feature = "thread_rng")]
 pub fn roll(input: &str) -> EResult {
-    Ok(RollBuilder::new().parse(input)?.into_roll()?.roll()?)
+    Ok(RollBuilder::new().parse(input)?.into_roll().unwrap().roll()?)
 }
 
 fn try_roll_expr_iter_with<I, R>(rng: &mut R, input: I) -> EResult
@@ -134,9 +157,9 @@ where
                 let res_val = res.value();
                 pairs.push((x, res));
                 total = total.checked_add(res_val).ok_or(if res_val > 0 {
-                    Error::OverflowPositive
+                    Error::OverflowPositive(OverflowPositive)
                 } else {
-                    Error::OverflowNegative
+                    Error::OverflowNegative(OverflowNegative)
                 })?;
             }
             Err(x) => return Err(x),
@@ -160,7 +183,7 @@ where
 
 #[cfg(test)]
 mod tests {
-    use crate::{roll, Die};
+    use crate::{roll, DiceTerm};
     #[test]
     fn arithmetic() {
         assert_eq!(roll("5 + 3").unwrap().total(), 8);
@@ -168,13 +191,9 @@ mod tests {
     }
     #[test]
     fn dice() {
-        let mut good = true;
-        match Die::new(0, 0) {
-            Ok(_) => good = false,
+        match DiceTerm::new(0, 0) {
+            Ok(_) => panic!(),
             Err(_) => (),
-        }
-        if !good {
-            panic!()
         }
     }
 }