+++ src/parse.rs
@@ -1,5 +1,5 @@
+//! Types and parsers for dice expressions.
 use crate::post::FormatOptions;
-use crate::Error;
 use nom::{
     branch::alt,
     bytes::complete::{tag, take_while1},
@@ -23,7 +23,7 @@ pub enum ParseError {
 }
 
 #[derive(Debug, Copy, Clone)]
-pub(crate) struct Die {
+pub struct DiceTerm {
     /// Negative numbers of dice are
     /// incorrect, but matching integer
     /// sizes is helpful.
@@ -31,38 +31,61 @@ pub(crate) struct Die {
     /// Negative dice sizes are nonsense,
     /// but matching integer sizes are helpful.
     pub(crate) size: i64,
+    // In particular, a proof we present in
+    // `crate::eval_term_with` is only valid
+    // due to our storing these things as
+    // signed integer types,
+    // despite their always being positive.
 }
-impl Die {
+impl DiceTerm {
     /// Creation of a `Die` may fail if:
     ///  - number of sides < 1
     ///  - number of dice  < 0
-    pub(crate) fn new(number: i64, size: i64) -> Result<Die, Error> {
+    pub(crate) fn new(number: i64, size: i64) -> Result<Self, InvalidDie> {
         // Forbid d0 and below. d1 is weird, but it
         // has a correct interpretation.
         if size < 1 || number < 0 {
-            Err(Error::InvalidDie)
+            Err(InvalidDie)
         } else {
-            Ok(Die { number, size })
+            Ok(DiceTerm { number, size })
         }
+    }
+    pub fn count(&self) -> u64 {
+        self.number as _
+    }
+    pub fn sides(&self) -> u64 {
+        self.size as _
     }
 }
 
 #[derive(Debug, Copy, Clone)]
-pub(crate) enum Term {
-    Die(Die),
+pub(crate) struct ConstantTerm {
+    /// This is not allowed to exceed what would be the u63 max,
+    /// and is not allowed to be less than zero.
+    value: i64,
+}
+impl Display for ConstantTerm {
+    fn fmt(&self, f: &mut std::fmt::Formatter) -> std::fmt::Result {
+        write!(f, "{}", self.value)
+    }
+}
+
+#[derive(Debug, Copy, Clone)]
+pub enum Term {
+    Dice(DiceTerm),
     Constant(i64),
 }
 impl Display for Term {
     fn fmt(&self, f: &mut Formatter) -> std::fmt::Result {
         match self {
-            Term::Die(x) => write!(f, "{}d{}", x.number, x.size),
+            Term::Dice(x) => write!(f, "{}d{}", x.number, x.size),
             Term::Constant(x) => write!(f, "{}", x),
         }
     }
 }
 
 #[derive(Debug, Clone, Copy)]
-pub(crate) enum Sign {
+pub enum Sign {
     Positive,
     Negative,
 }
@@ -126,13 +149,84 @@ impl Display for Expr {
     }
 }
 
-pub(crate) type Expression = Vec<Expr>;
+#[derive(Debug)]
+pub struct Expression {
+    exprs: Vec<Expr>,
+}
+impl Expression {
+    pub(crate) fn new(exprs: Vec<Expr>) -> Self {
+        Expression { exprs }
+    }
+    pub(crate) fn iter(&self) -> ExpressionRefIterator<'_> {
+        ExpressionRefIterator {
+            internal_iterator: self.exprs.iter(),
+        }
+    }
+    // This could be a trait implementation, but it's not supposed
+    // to be visible outside of this crate.
+    pub(crate) fn into_iter(self) -> ExpressionIterator {
+        ExpressionIterator {
+            internal_iterator: self.exprs.into_iter(),
+        }
+    }
+    pub fn terms(&self) -> TermIter {
+        TermIter { internal_iterator: self.iter() }
+    }
+    pub fn roll_with<R: ::rand::Rng>(&self, rng: &mut R) -> Result<crate::ExpressionResult, crate::Error> {
+        crate::roll_expr_iter_with(rng, self.iter().copied())
+    }
+    #[cfg(feature = "thread_rng")]
+    pub fn roll(&self) -> crate::EResult {
+        self.roll_with(&mut ::rand::thread_rng())
+    }
+    /// Nom parser for an `Expression`.
+    ///
+    /// This is the same as `parse::dice`,
+    /// provided here as well to help discoverability.
+    pub fn parse(input: &str) -> IResult<&str, Result<Self, InvalidDie>> {
+        dice(input)
+    }
+}
+pub(crate) struct ExpressionRefIterator<'a> {
+    internal_iterator: ::std::slice::Iter<'a, Expr>,
+}
+impl<'a> Iterator for ExpressionRefIterator<'a> {
+    type Item = &'a Expr;
+    fn next(&mut self) -> Option<Self::Item> {
+        self.internal_iterator.next()
+    }
+}
+pub(crate) struct ExpressionIterator {
+    internal_iterator: ::std::vec::IntoIter<Expr>,
+}
+impl Iterator for ExpressionIterator {
+    type Item = Expr;
+    fn next(&mut self) -> Option<Self::Item> {
+        self.internal_iterator.next()
+    }
+}
+pub struct TermIter<'a> {
+    internal_iterator: ExpressionRefIterator<'a>,
+}
+impl<'a> Iterator for TermIter<'a> {
+    type Item = &'a Term;
+    fn next(&mut self) -> Option<Self::Item> {
+        self.internal_iterator.next().map(|x| &x.term)
+    }
+}
 
 fn is_dec_digit(c: char) -> bool {
     c.is_digit(10)
 }
 
-fn integer(input: &str) -> IResult<&str, i64> {
+/// Parser for an effectively 63-bit unsigned integer.
+///
+/// This is useful because we represent signs separately
+/// from the dice they might be seen as attached to.
+/// This is due to a semantic difference.
+/// e.g., In the real world, you don't have negative four dice.
+/// You subtract the result of four dice.
+pub fn integer(input: &str) -> IResult<&str, i64> {
     let (input, int) = take_while1(is_dec_digit)(input)?;
     // Pretend to be a 63 bit unsigned integer.
     let i = match int.parse::<i64>() {
@@ -145,13 +239,46 @@ fn integer(input: &str) -> IResult<&str, i64> {
     Ok((input, i))
 }
 
-fn die(input: &str) -> IResult<&str, Term> {
+#[derive(Error, Debug)]
+#[error("invalid die")]
+pub struct InvalidDie;
+impl From<InvalidDie> for ParseError {
+    fn from(_: InvalidDie) -> Self {
+        Self::InvalidExpression
+    }
+}
+
+type PResult<I, O, PE, E = (I, ::nom::error::ErrorKind)> = Result<(I, Result<O, PE>), ::nom::Err<E>>;
+fn okay<I, T, PE, E>(input: I, exp: T) -> Result<(I, Result<T, PE>), E> {
+    Ok((input, Ok(exp)))
+}
+fn purr<I, T, PE, E>(input: I, err: PE) -> Result<(I, Result<T, PE>), E> {
+    Ok((input, Err(err)))
+}
+macro_rules! trip {
+    ($in:expr, $exp:expr) => {
+        match $exp {
+            Ok(x) => x,
+            Err(e) => return Ok(($in, Err(e))),
+        }
+    }
+}
+fn die(input: &str) -> PResult<&str, DiceTerm, InvalidDie> {
     // number of dice : [integer]
     // separator      : "d"
     // size of dice   : integer
     let (input, (number, _, size)) = tuple((opt(integer), tag("d"), integer))(input)?;
     let number = number.unwrap_or(1);
-    Ok((input, Term::Die(Die { number, size })))
+    // Note that since we use the bare DiceTerm constructor,
+    // we need to make certain no invalid dice are created.
+    // That means `number` needs to be >= 0, and `size` needs to be >= 1.
+    // Given that `integer` does not create integers less than zero,
+    // the only check we need to do here is `size != 0`.
+    if size != 0 {
+        okay(input, DiceTerm { number, size })
+    } else {
+        purr(input, InvalidDie)
+    }
 }
 
 fn addition(input: &str) -> IResult<&str, Sign> {
@@ -167,18 +294,28 @@ fn operator(input: &str) -> IResult<&str, Sign> {
     alt((addition, subtraction))(input)
 }
 
-fn whitespace(input: &str) -> IResult<&str, &str> {
+/// Parser for a `+` or `-` sign.
+pub fn sign(input: &str) -> IResult<&str, Sign> {
+    // While currently equivalent to `operator`,
+    // the idea of a sign will never change, whilst `operator`
+    // may be extended to recognize other operators,
+    // which would break someone who just wants
+    // a positive or negative sign.
+    // TL;DR: Arithmetic != Sign
+    alt((addition, subtraction))(input)
+}
+
+/// Nom parser for whitespace
+pub fn whitespace(input: &str) -> IResult<&str, &str> {
     alt((tag(" "), tag("\t")))(input)
 }
 
 fn separator(input: &str) -> IResult<&str, Sign> {
-    let (input, t) = tuple((many0(whitespace), operator, many0(whitespace)))(input)?;
-    Ok((input, t.1))
+    tuple((many0(whitespace), operator, many0(whitespace)))(input).map(|(i, (_, op, _))| (i, op))
 }
 
-fn constant(input: &str) -> IResult<&str, Term> {
-    let i = integer(input)?;
-    Ok((i.0, Term::Constant(i.1)))
+fn constant(input: &str) -> IResult<&str, ConstantTerm> {
+    integer(input).map(|(i, int)| (i, ConstantTerm { value: int }))
 }
 
 // /// Use like this, where map is a HashMap: `|x| variable(map, x)`
@@ -188,20 +325,25 @@ fn constant(input: &str) -> IResult<&str, Term> {
 //     Ok((input, Term::Constant(v)))
 // }
 
-fn term(input: &str) -> IResult<&str, Term> {
-    alt((die, constant))(input)
+fn term(input: &str) -> PResult<&str, Term, InvalidDie> {
+    alt((
+        |x| die(x).map(|(i, d)| (i, d.map(Term::Dice))),
+        |x| constant(x).map(|(i, c)| (i, Ok(Term::Constant(c.value)))),
+    ))(input)
 }
 
-fn dice(input: &str) -> IResult<&str, Expression> {
-    // [(+/-)] die ((+/-) die)*
+/// Nom parser for a dice expression.
+pub fn dice(input: &str) -> PResult<&str, Expression, InvalidDie> {
+    // [(+/-)] dice ((+/-) dice)*
     let (input, (sign, term, terms)) =
         tuple((opt(separator), term, many0(tuple((separator, term)))))(input)?;
     let sign = sign.unwrap_or(Sign::Positive);
+    let term = trip!(input, term);
     let mut expression = vec![Expr { term, sign }];
     for (sign, term) in terms {
-        expression.push(Expr { term, sign });
+        expression.push(Expr { term: trip!(input, term), sign })
     }
-    Ok((input, expression))
+    okay(input, Expression::new(expression))
 }
 
 /// Wrap up getting errors from parsing a dice expression.
@@ -214,6 +356,6 @@ pub(crate) fn wrap_dice(input: &str) -> Result<Expression, ParseError> {
     if !input.is_empty() {
         Err(ParseError::InvalidExpression)
     } else {
-        Ok(e)
+        e.map_err(|e| e.into())
     }
 }